Adds a test workflow with Github Action
This is based on experimentation and
https://help.github.com/en/actions/language-and-framework-guides/using-python-with-github-actions,
there may be other ways to do this.

This workflow will run everytime a commit is pushed to a branch and when
a Pull Request is opened for that branch. Just running `python
manage.py test` checks that no syntax errors exist, and will also run
all of the tests we put into `tests.py`.