Add setting to run Cloud SQL Proxy locally
To use SQLite database locally, keep the settings as is.

To use the Cloud SQL Proxy which is a local MySQL database that connects
to the Google Cloud MySQL database, set DEBUG=False and follow the
directions in the comments.

Before deploying the code to Google Cloud, set DEBUG=False and the rest
should take care of itself