Track migrations __init__ in git
After much googling and StackOverflow posts, discovered that Django for some reason won't create the migrations using `python
manage.py makemigrations` if the "migrations" folder doesn't exist
within the application with a file named "__init__.py"

This is the reason why we had to run "python manage.py migate --sync-db"
which is discouraged by the documentation. With the __init__.py file it
should work correctly.