Remove STATIC_ROOT setting
According to Django docs, we were supposed to put static files (CSS and
JS) in a "static" folder within each app, not at the global level. Then,
running `python manage.py collectstatic` would take the static files
from each app and put them into the global 'static' folder *at deploy
time*, so that it works in both development and production.

Since we haven't been doing it the right way and it would be a lot of
work to change, removing this setting so that it continues to work the
way we've been doing it.

See https://docs.djangoproject.com/en/3.0/howto/static-files/ for more
details.